requirements.txt
pynacl
httpx
python-fasthtml>=0.4.4

